+++ src/Himari/SafePrelude.hs
@@ -1,0 +1,60 @@
+{-# LANGUAGE Safe #-}
+{-# LANGUAGE NoGeneralizedNewtypeDeriving #-}
+
+-- | Subset of "Himari.Prelude" that can be imported under Safe Haskell.
+-- Re-exported by "Himari.Prelude" so most users do not need to import this directly.
+module Himari.SafePrelude
+  ( module Export
+  ) where
+
+import Control.Applicative as Export
+import Control.DeepSeq as Export
+import Control.Lens as Export
+import Control.Monad as Export
+import Control.Monad.Cont as Export
+import Control.Monad.Logger as Export
+import Control.Monad.Reader as Export
+import Control.Monad.ST as Export
+import Control.Monad.State.Strict as Export
+import Control.Monad.Writer.CPS as Export
+import Data.Bifoldable as Export
+import Data.Bitraversable as Export
+import Data.Bool as Export
+import Data.Complex as Export
+import Data.Convertible as Export
+import Data.Default as Export
+import Data.Either as Export
+import Data.Eq as Export
+import Data.Fixed as Export
+import Data.Foldable as Export
+import Data.Function as Export
+import Data.Functor.Compose as Export
+import Data.Hashable as Export
+import Data.Int as Export
+import Data.Kind as Export
+import Data.Maybe as Export
+import Data.Ord as Export
+import Data.Proxy as Export
+import Data.Ratio as Export
+import Data.String as Export
+import Data.Time as Export
+import Data.Traversable as Export
+import Data.Tuple as Export
+import Data.Void as Export
+import Data.Word as Export
+import GHC.Stack as Export
+import Himari.Prelude.Arrow as Export
+import Himari.Prelude.Catch as Export
+import Himari.Prelude.Category as Export
+import Himari.Prelude.Data as Export
+import Himari.Prelude.FilePath as Export
+import Himari.Prelude.Generics as Export
+import Himari.Prelude.Monoid as Export
+import Himari.Prelude.Safe as Export
+import Himari.Prelude.Type as Export
+import Numeric as Export
+import Numeric.Natural as Export
+import Text.Pretty.Simple as Export
+import Text.Show as Export
+import UnliftIO.Environment as Export
+import Prelude as Export